IIMM-main/requirements.txt
torch>=2.0.1
numpy>=1.24.3
pandas>=1.5.3
scipy>=1.10.0
scikit-learn>=1.2.1
wilds>=2.0.0
open_clip_torch>=2.24.0
transformers==4.54.0
